+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to `gent/services-opening-hours` package.
 
-## [Unreleased]
+## [1.1.0]
 
 ### Changed
 
@@ -58,6 +58,8 @@ hours for a Channel.
 * DMOH-17: Added the OpeningHoursPeriod & OpeningHoursPeriodHTML methods to the
   ChannelService.
 
+[1.1.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/1.0.1...1.1.0
+[1.0.1]: https://github.com/StadGent/php_package_services-opening-hours/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/0.2.0...1.0.0
 [0.2.0]: https://github.com/StadGent/php_package_services-opening-hours/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/StadGent/php_package_services-opening-hours/releases/tag/0.1.0